requirements.txt
selenium
pillow
reportlab